Moved negHex from aarch64 specific normalization to StringUtils
