+++ fleece/decoders/aarch64_common.C
@@ -78,25 +78,6 @@ void removeExtraZeroesFromFmovImm(char* buf, int bufLen) {
 
 }
 
-
-char negHex(char h) {
-   
-   if (h >= '0' && h <= '5') {
-      return 'f' + '0' - h;
-   }
-
-   if (h >= '6' && h <= '9') {
-      return '9' + '6' - h;
-   }
-
-   if (h >= 'a' && h <= 'f') {
-      return 'f' + '0' - h;
-   }
-
-   return h;
-}
-
-
 void aliasMovn(char* buf, int bufLen) {
    if (strncmp(buf, "movn", 4)) {
       return;
